Fix for accessing sandbox RV without proxy
When PRI components (to0client and device) are used in open network,
they are not able to connect to hosted Rendezvous Service.

The issue is resolved by setting the HTTP version to 1.1 explicitly
while creating the required HTTP clients.

While at it, update the third-party package vesions to fix known
security vulnerabilities.

Signed-off-by: Behera, Tushar Ranjan <tushar.ranjan.behera@intel.com>